pyproject.toml
[build-system]
requires = [
    "wheel",
    "setuptools>=61.0.0",
    "oldest-supported-numpy",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "qseek"
requires-python = ">=3.10,<3.13"
license = { text = "GPLv3" }
dynamic = ["version"]

description = "qseek - The earthquake detector."
readme = "README.md"
authors = [
    { name = "Marius Paul Isken", email = "mi@gfz-potsdam.de" },
    { name = "Sebastian Heimann", email = "heimann@uni-potsdam.de" },
    { name = "Marius Kriegerowski", email = "mk@quakesaver.de" },
]
maintainers = [{ name = "Marius Paul Isken", email = "mi@gfz-potsdam.de" }]
keywords = [
    "seismology",
    "earthquake",
    "seismicity",
    "detection",
    "localization",
    "stacking-migration",
    "beamforming",
]

dependencies = [
    "numpy>=1.17.3, <2",
    "scipy>=1.8.0",
    "pyrocko>=2022.06.10",
    "seisbench>=0.5.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.8",
    "aiohttp_cors>=0.7.0",
    "typing-extensions>=4.6",
    "lru-dict>=1.3",
    "rich>=13.4",
    "nest_asyncio>=1.5",
    "pyevtk>=1.6",
    "psutil>=5.9",
    "aiofiles>=23.0",
    "typer >=0.12.3",
    "scikit-fmm >= 2024.05",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Typing :: Typed",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]

[project.optional-dependencies]
dev = ["pre-commit>=3.4", "ruff>=0.3.0", "pytest>=7.4", "pytest-asyncio>=0.21"]

docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "griffe-pydantic>=1.1.0",
    "markdown-exec>=1.9",
]
completion = ["argcomplete>=3.2"]

[project.scripts]
qseek = "qseek.apps.qseek:main"

[tool.setuptools.packages.find]
where = ["src/"]
include = ["qseek"]

[project.urls]
Home = "https://pyrocko.org"
GitHub = "https://github.com/pyrocko/qseek"
Issues = "https://github.com/pyrocko/qseek/issues"

[tool.setuptools_scm]

[tool.ruff.lint]
extend-select = [
    'W',
    'N',
    'DTZ',
    'FA',
    'G',
    'RET',
    'SIM',
    'B',
    'RET',
    'C4',
    'I',
    'RUF',
    'T20',
    'D',
]

ignore = [
    "RUF012",
    "RUF009",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D107",
]

[tool.ruff]
target-version = 'py311'

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
markers = ["plot: plot figures in tests"]

[tool.cibuildwheel.macos]
before-build = "brew install libomp"
